fix recordbag selection render
