+++ pages/_app.js
@@ -1,28 +1,15 @@
 import GlobalStyle from '../components/GlobalStyle';
 import Layout from '../components/Layout';
-import path from 'path';
-import fsPromises from 'fs/promises';
 import { useState } from 'react';
+import collection from '../json/data-dummy.json';
 
-export async function getStaticProps() {
-  const filePath = path.join(process.cwd(), './json/data-dummy.json');
-  const jsonData = await fsPromises.readFile(filePath);
-  const collectionData = JSON.parse(jsonData);
-
-  return {
-    props: collectionData,
-  };
-}
-
-function MyApp({ Component, pageProps, props }) {
-  const collection = props.collection;
-
-  const myCollection = collection.map((file) => {
-    return { ...file, isChecked: false };
-  });
-
-  const [collectionState, setCollectionState] = useState(myCollection);
-  const [selection, setSelection] = useState(selectedRecords);
+function MyApp({ Component, pageProps }) {
+  const [collectionState, setCollectionState] = useState(() =>
+    collection.map((file) => {
+      return { ...file, isChecked: false };
+    })
+  );
+  console.log(collectionState[0], 'myApp');
 
   return (
     <>
@@ -31,9 +18,7 @@ function MyApp({ Component, pageProps, props }) {
         <Component
           {...pageProps}
           collectionState={collectionState}
-          setCollectionState={setCollectionState}
-          selection={selection}
-          setSelection={setSelection}
+          onSetCollectionState={setCollectionState}
         />
       </Layout>
     </>
